pyproject.toml
[tool.poetry]
name = "harmo"
version = "1.0.4"
description = "Harmony（和谐）的缩写"
authors = ["mongnet <250021520@qq.com>"]
license = "MIT"
readme = ["README.md"]
homepage = "https://github.com/mongnet/harmo"
repository = "https://github.com/mongnet/harmo"
keywords = ["api", "swagger", "pytest", "harmo"]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"
faker = "^14.2.1"
chevron = "^0.14.0"
py = "^1.11.0"
pathlib2 = "^2.3.7"
xmltodict = "^0.13.0"
suds-py3 = "^1.4.5"
"ruamel.yaml" = "^0.17.21"
cleo = "^2.0.1"
openpyxl = "^3.1.2"
mitmproxy = "^9.0.1"
pytest-metadata = "3.0.0"
jinja2 = "^3.1.2"
jmespath = "^1.0.1"
pytest-html = "^4.1.1"
pytest = "^8.2.0"
requests = "^2.31.0"
jsonpath = "^0.82.2"
allure-pytest = "^2.13.5"
beautifulsoup4 = "^4.12.2"
playwright = "^1.40.0"
pytest-playwright = "^0.4.4"
sqlalchemy = "^2.0.23"
jsonschema = "^4.20.0"
genson = "^1.2.2"
pyjwt = "^2.8.0"

[tool.poetry.dev-dependencies]
pytest = "^8.2"

[tool.poetry.scripts]
harmo = "harmo.console:main"

[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[tool.poetry.plugins."pytest11"]
"harmo_plugin" = "harmo.plugin"
"harmo_fixture" = "harmo.fixture"

[build-system]
requires = ["poetry>=1.8"]
build-backend = "poetry.masonry.api"

